backend/requirements.txt
fastapi
uvicorn[standard]
python-docx
pyyaml
requests
python-dotenv
pydantic-settings
openai
